// projects/app-common/src/lib/appCommon/models/enum/FilterOperationEnum.ts
export enum FilterOperationEnum {
    GREATER_THAN = 'gt',
    LESS_THAN = 'lt',
    GREATER_THAN_EQUAL = 'gte',
    LESS_THAN_EQUAL = 'lte',
    NOT_EQUAL = 'not_equal',
    EQUAL = 'equal',
    MATCH = 'match',
    MATCH_START = 'match_start',
    MATCH_END = 'match_end',
    IN_LONG_LIST = 'in_Long_List',
    NOT_IN_LONG_LIST = 'not_in_Long_List',
    BIT_CHECK = 'bit_check',
    DAY_EQUAL = 'day_equal'
}